# Conflicts:
#	central-bank-rest-src/src/main/java/ru/bluewater/centralbankrestsrc/controller/BICDirectoryEntryControllerImpl.java
#	central-bank-rest-src/src/main/java/ru/bluewater/centralbankrestsrc/entity/ED807Entity.java
#	central-bank-rest-src/src/main/java/ru/bluewater/centralbankrestsrc/mapper/AccRstrListEntityMapper.java
#	central-bank-rest-src/src/main/java/ru/bluewater/centralbankrestsrc/mapper/BICDirectoryEntryEntityMapper.java
#	central-bank-rest-src/src/main/java/ru/bluewater/centralbankrestsrc/service/BICDirectoryEntryService.java

+++ central-bank-rest-api/src/main/java/ru/bluewater/centralbankrestapi/api/dto/request/PartInfoRequestDTO.java
@@ -3,6 +3,7 @@ package ru.bluewater.centralbankrestapi.api.dto.request;
 import lombok.*;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.UUID;
 @Getter
 @Setter
@@ -10,5 +11,5 @@ public class PartInfoRequestDTO {
     private UUID uuid;
     private Integer partNo;
     private Integer partQuantity;
-    private BigDecimal partAggregateID;
+    private BigInteger partAggregateID;
 }
